pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nah"
version = "0.1.0"
description = "Context-aware safety guard for Claude Code. A permission system you control."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Manuel Schipper" },
]
keywords = ["claude", "claude-code", "safety", "guard", "hooks", "security", "permissions"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: Software Development :: Quality Assurance",
]

[project.optional-dependencies]
dev = ["pytest"]
config = ["pyyaml>=6.0"]

[project.scripts]
nah = "nah.cli:main"

[project.urls]
Homepage = "https://github.com/manuelschipper/nah"
Issues = "https://github.com/manuelschipper/nah/issues"

[tool.pytest.ini_options]
testpaths = ["tests"]
